XRDDEV-1017: Make SignerClient more robust
In certain conditions, SignerClient can observe timeout when Signer
is restarted while a request is being made. Make SignerClient
to watch Signer's RequestProcessor actor, and in case of termination,
try to re-establish connection as soon as possible. Alter the Signer
startup to remove unnecessary delays.

Also update Akka to the latest patch version (2.6.6).